cmd: add command for generating keypairs
See issue golang/go#33923, I had mistakenly used `openssl rand -hex
64` to generate a private key, which did not work for a pretty subtle
reason - usually the data you are attempting to sign has been subtly
changed between when you signed the message and when you verified it
but here the key was the problem.

Attempt to fix this by making it really easy for users to generate
keys for use with this library.